chore: move gql to config
